Compute weighted expectations

The E_loo() function computes weighted expectations (means, variances, quantiles) using the importance weights obtained from the PSIS smoothing procedure. The expectations estimated by the E_loo() function assume that the PSIS approximation is working well. A small Pareto k estimate is necessary, but not sufficient, for E_loo() to give reliable estimates. Additional diagnostic checks for gauging the reliability of the estimates are in development and will be added in a future release.

E_loo(x, psis_object, ...)

# S3 method for default
E_loo(
  x,
  psis_object,
  ...,
  type = c("mean", "variance", "quantile"),
  probs = NULL,
  log_ratios = NULL
)

# S3 method for matrix
E_loo(
  x,
  psis_object,
  ...,
  type = c("mean", "variance", "quantile"),
  probs = NULL,
  log_ratios = NULL
)

Arguments

x
A numeric vector or matrix.

psis_object
An object returned by psis().

...
Arguments passed to individual methods.

type
The type of expectation to compute. The options are "mean", "variance", and "quantile".

probs
For computing quantiles, a vector of probabilities.

log_ratios
Optionally, a vector or matrix (the same dimensions as x) of raw (not smoothed) log ratios. If working with log-likelihood values, the log ratios are the negative of those values. If log_ratios is specified we are able to compute Pareto k diagnostics specific to E_loo().

Value

A named list with the following components:

value

The result of the computation.For the matrix method, value is a vector with ncol(x) elements, with one exception: when type="quantile" and multiple values are specified in probs the value component of the returned object is a length(probs) by ncol(x) matrix.For the default/vector method the value component is scalar, with one exception: when type is "quantile" and multiple values are specified in probs the value component is a vector with length(probs) elements.

pareto_k

Function-specific diagnostic.If log_ratios is not specified when calling E_loo(), pareto_k will be NULL. Otherwise, for the matrix method it will be a vector of length ncol(x) containing estimates of the shape parameter \(k\) of the generalized Pareto distribution. For the default/vector method, the estimate is a scalar.

Examples

# \donttest{
# Use rstanarm package to quickly fit a model and get both a log-likelihood
# matrix and draws from the posterior predictive distribution
library("rstanarm")

#> Loading required package: Rcpp

#> Registered S3 method overwritten by 'xts':
#>   method     from
#>   as.zoo.xts zoo 

#> rstanarm (Version 2.19.3, packaged: )

#> - Do not expect the default priors to remain the same in future rstanarm versions.

#> Thus, R scripts should specify priors explicitly, even if they are just the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#> options(mc.cores = parallel::detectCores())

#> - bayesplot theme set to bayesplot::theme_default()

#>    * Does _not_ affect other ggplot2 plots

#>    * See ?bayesplot_theme_set for details on theme setting

# data from help("lm")
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
d <- data.frame(
  weight = c(ctl, trt),
  group = gl(2, 10, 20, labels = c("Ctl","Trt"))
)
fit <- stan_glm(weight ~ group, data = d, refresh = 0)
yrep <- posterior_predict(fit)
dim(yrep)

#> [1] 4000   20

log_ratios <- -1 * log_lik(fit)
dim(log_ratios)

#> [1] 4000   20

r_eff <- relative_eff(exp(-log_ratios), chain_id = rep(1:4, each = 1000))
psis_object <- psis(log_ratios, r_eff = r_eff, cores = 2)

E_loo(yrep, psis_object, type = "mean")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.125699 4.957090 5.027723 4.901066 5.077694 5.067100 5.005037 5.101876
#>  [9] 5.001453 5.008510 4.630159 4.737338 4.695085 4.776057 4.527644 4.741270
#> [17] 4.511924 4.645113 4.712650 4.648726
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "var")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 0.0001758090 0.0001929221 0.0002846390 0.0002322068 0.0002045981
#>  [6] 0.0002301491 0.0003099407 0.0002048676 0.0002484030 0.0003357314
#> [11] 0.0003004722 0.0001929832 0.0002825810 0.0001927456 0.0002933647
#> [16] 0.0001840801 0.0003831361 0.0002629442 0.0002561911 0.0003289782
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = 0.5) # median

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.137205 4.949528 5.024313 4.907747 5.075786 5.064358 5.010031 5.116541
#>  [9] 4.996308 5.009588 4.623399 4.731390 4.696351 4.756020 4.523384 4.735937
#> [17] 4.521804 4.650145 4.714856 4.639187
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = c(0.1, 0.9))

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 4.205287 3.959167 4.057398 3.973357 4.095659 4.096927 3.989554 4.124568
#> [2,] 6.039962 5.957049 5.989300 5.824918 6.059539 6.068418 6.015977 6.092703
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]   [,16]
#> [1,] 4.019989 3.970864 3.651567 3.756684 3.709367 3.886955 3.597137 3.79381
#> [2,] 5.980762 6.049186 5.608002 5.734082 5.682718 5.688905 5.446165 5.70192
#>         [,17]    [,18]    [,19]    [,20]
#> [1,] 3.651192 3.686134 3.710400 3.666609
#> [2,] 5.377819 5.637251 5.749822 5.648830
#> 
#> $pareto_k
#> NULL
#> 

# To get Pareto k diagnostic with E_loo we also need to provide the negative
# log-likelihood values using the log_ratios argument.
E_loo(yrep, psis_object, type = "mean", log_ratios = log_ratios)

#> $value
#>  [1] 5.125699 4.957090 5.027723 4.901066 5.077694 5.067100 5.005037 5.101876
#>  [9] 5.001453 5.008510 4.630159 4.737338 4.695085 4.776057 4.527644 4.741270
#> [17] 4.511924 4.645113 4.712650 4.648726
#> 
#> $pareto_k
#>  [1]  0.310077357  0.119040882  0.027743035  0.308682529  0.145106471
#>  [6]  0.078710510  0.158699038  0.044389923 -0.051546216  0.048681679
#> [11]  0.088539247 -0.005486665  0.166370883  0.274525487  0.200213915
#> [16]  0.178168756  0.230236697  0.082111462  0.083460675  0.149581692
#> 

# }