Model averaging via stacking or Pseudo-BMA weighting.

Model averaging via stacking or Pseudo-BMA weighting.

model_weights(log_lik_list, method = "stacking", BB = T, BB_n = 1000,
  alpha = 1, seed = NULL, optim_method = "BFGS")

Arguments

log_lik_list
A list of pointwise log likelihood simulation matrixes. The \(i\) th element corresponds to the \(i\) th model. Each row of the matrix is the log likelihood vector evaluated using a simulated parameter

method
One of "stacking" or "pseudobma", indicating which method is to use for obtaining the optimal weights. "stacking" refers to stacking of predictive distributions and "pseudobma" refers to Pseudo-BMA weighting (by setting "BB"=F) or Pseudo-BMA+ weighting (by setting "BB"=T).

BB
Logicals used when "method"="pseudobma". If True(default), Bayesian Bootstrap will be used to adjust the Pseudo-BMA weighting, which is called Pseudo-BMA+ weighting. It helps regularize the weight away from 0 and 1, so as to reduce the variance.

BB_n
A positive integer indicating the number of samples in Bayesian Bootstrap. It is necessary when BB=TRUE. The default number is 1000.

alpha
A positive scalar; the shape parameter in the Dirichlet distribution when doing Bootstrap. The default is \(1\).

seed
An integer; optional. If specified, it will fix the random seed when dong Bayesian Bootstrap sampling.

optim_method
The optimization method to be used in stacking. It can be chosen from "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent". The default method is "BFGS".

Value

A vector of optimal model weights.

Details

This function implements stacking of predictive distributions, Pseudo-BMA and Pseudo-BMA+ weighting for combining multiple predictive distributions.

For either method, we can use Leave-one-out cross-validation (LOO) to estimate the expected log predictive density(elpd). Stacking combines all model by maximizing the leave-one-out predictive density of the combination distribution. Pseudo-BMA finds the relative weights proportional to elpd of each model. Pseudo-BMA+ takes into account the uncertainty resulted from having a finite number of proxy samples from the future data distribution through Bayesian bootstrap (set "BB=T"), which will keep weights further away from 0 and 1.

In general, we recommend stacking for averaging predictive distributions, while Pseudo-BMA+ can serve as a computationally easier alternative.

See also

model_select for single-model selection. pseudobma_weight for details on Pseudo-BMA and Pseudo-BMA+ weights. stacking_weight for details on stacking weighs. constrOptim for the choice of optimization methods.

Examples

not_run({
  log_lik1 <- extract(stan(model=model_1,data=data))[['log_lik']]
  log_lik2 <- extract(stan(model=model_1,data=data))[['log_lik']]
  w1=model_weights(list(log_lik1, log_lik2),method="stacking")
  w2=model_weights(list(log_lik1, log_lik2),method="pseudobma",BB=T)
})