Named lists

Create a named list using specified names or, if names are omitted, using the names of the objects in the list. The code list(a = a, b = b) becomes nlist(a,b) and list(a = a, b = 2) becomes nlist(a, b = 2), etc.

nlist(...)

Arguments

...
Objects to include in the list.

Value

A named list.

Examples

# All variables already defined
a <- rnorm(100)
b <- mat.or.vec(10, 3)
nlist(a,b)

#> $a
#>   [1]  1.10811958 -0.87829676 -1.65681593 -1.69063570 -0.27132048 -1.15584660
#>   [7]  0.20405452  0.04701523 -0.51550139  0.01153985  0.55204616  1.32149314
#>  [13]  1.94724115 -0.57149695 -0.69384290 -1.96580236 -0.08385387  0.27190962
#>  [19] -2.46672879 -1.88824979  0.76100617 -1.31066493 -0.50296899  0.09770720
#>  [25]  0.69068959  1.30879828  0.46884941 -0.83426786  1.10167012 -1.48977164
#>  [31] -1.05975718 -0.24901157  0.64253085 -1.13949482 -0.53850692 -0.85775871
#>  [37] -0.73824100  0.60811926 -2.00208705 -0.56113495  0.42204447  0.56755715
#>  [43]  1.18565339 -1.19922441  0.52796943  2.57753602  0.19146928 -0.12950747
#>  [49] -0.47049674 -0.47124834 -0.45861253  0.88321666  2.17935234  0.78069654
#>  [55] -0.38664165  0.11466462  0.14008793  0.78055139 -0.63465685 -0.79332921
#>  [61] -0.85237504  0.55025138 -0.03166208 -1.80540296  0.37736434 -1.01891590
#>  [67] -0.45123629  1.03551642 -2.98295122 -1.34937026  1.50652682  1.33540584
#>  [73]  0.67021191 -0.97962549  1.07984518  1.49631601  0.02567140  0.07393076
#>  [79]  0.08410629 -0.54413205 -2.01570102 -1.57668200 -0.49772961 -0.28488772
#>  [85] -0.32997572  0.74003464  1.03634277  1.78561146  0.82658104  1.75847216
#>  [91]  2.00654139  0.17887677 -0.19350366  1.24979887 -1.58308710 -1.78250532
#>  [97]  0.58171191  0.27160900 -1.68228269 -1.53966793
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 

# Define some variables in the call and take the rest from the environment
nlist(a, b, veggies = c("lettuce", "spinach"), fruits = c("banana", "papaya"))

#> $a
#>   [1]  1.10811958 -0.87829676 -1.65681593 -1.69063570 -0.27132048 -1.15584660
#>   [7]  0.20405452  0.04701523 -0.51550139  0.01153985  0.55204616  1.32149314
#>  [13]  1.94724115 -0.57149695 -0.69384290 -1.96580236 -0.08385387  0.27190962
#>  [19] -2.46672879 -1.88824979  0.76100617 -1.31066493 -0.50296899  0.09770720
#>  [25]  0.69068959  1.30879828  0.46884941 -0.83426786  1.10167012 -1.48977164
#>  [31] -1.05975718 -0.24901157  0.64253085 -1.13949482 -0.53850692 -0.85775871
#>  [37] -0.73824100  0.60811926 -2.00208705 -0.56113495  0.42204447  0.56755715
#>  [43]  1.18565339 -1.19922441  0.52796943  2.57753602  0.19146928 -0.12950747
#>  [49] -0.47049674 -0.47124834 -0.45861253  0.88321666  2.17935234  0.78069654
#>  [55] -0.38664165  0.11466462  0.14008793  0.78055139 -0.63465685 -0.79332921
#>  [61] -0.85237504  0.55025138 -0.03166208 -1.80540296  0.37736434 -1.01891590
#>  [67] -0.45123629  1.03551642 -2.98295122 -1.34937026  1.50652682  1.33540584
#>  [73]  0.67021191 -0.97962549  1.07984518  1.49631601  0.02567140  0.07393076
#>  [79]  0.08410629 -0.54413205 -2.01570102 -1.57668200 -0.49772961 -0.28488772
#>  [85] -0.32997572  0.74003464  1.03634277  1.78561146  0.82658104  1.75847216
#>  [91]  2.00654139  0.17887677 -0.19350366  1.24979887 -1.58308710 -1.78250532
#>  [97]  0.58171191  0.27160900 -1.68228269 -1.53966793
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 
#> $veggies
#> [1] "lettuce" "spinach"
#> 
#> $fruits
#> [1] "banana" "papaya"
#>