Stacking of predictive distributions

Stacking of predictive distributions

stacking_weight(lpd_point, optim_method = "BFGS")

Arguments

lpd_point
A matrix of pointwise leave-one-out likelihood evaluated in different models. Each column corresponds to one model.

optim_method
The optimization method to be used in stacking; it can be chosen from "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent". The default one is "BFGS".

Value

A vector of best model weights that maximize the leave-one-out log score of the combination.

Details

lpd_point is a matrix of pointwise log leave-one-out likelihood, which can be calculated from loo or through running exact LOO in advance. It should be a \(N\) by \(K\) matrix when Sample size is \(N\) and model number is \(K\). stacking is an approach that finds the optimal linear combining weight which maximizes the leave-one-out log score.

See also

loo for details on leave-one-out elpd estimation. pseudobma_weight for model weighting by Pseudo-BMA and Bayesian bootstrap adjustment. optim for the choice of optimization methods.

Examples

not_run({
  loo1 <- loo(log_lik1)$pointwise[,1]
  loo2 <- loo(log_lik2)$pointwise[,1]
  stacking_weight(cbind(loo1, loo2))
})