Compute weighted expectations

The E_loo function computes weighted expectations (means, variances, quantiles) using the smoothed importance weights obtained from the PSIS procedure. The expectations estimated by the E_loo function assume that the PSIS approximation is working well. A small Pareto k estimate is necessary, but not sufficient, for E_loo to give reliable estimates. Additional diagnostic checks for gauging the reliability of the estimates are in development and will be added in a future release.

E_loo(x, lw, ...)

# S3 method for default
E_loo(x, lw, ..., type = c("mean", "var", "quantile"),
  probs)

# S3 method for matrix
E_loo(x, lw, ..., type = c("mean", "var", "quantile"), probs)

Arguments

x
A numeric matrix or vector.

lw
A numeric matrix (or vector) of smoothed log-weights with the same dimensions (or length) as x. Typically this will be the "lw_smooth" element of the list returned by psislw.

...
For the generic function, arguments passed to the individual methods.

type
The type of expectation to compute. The options are "mean", "var" (variance), and "quantile".

probs
A vector of probabilities. Ignored unless type is "quantile".

Value

The matrix method returns a vector with ncol(x) elements, with one exception: when type is "quantile" and multiple values are specified in probs the returned object is a length(probs) by ncol(x) matrix.

The default/vector method returns a scalar, with one exception: when type is "quantile" and multiple values are specified in probs the returned object is a vector with length(probs) elements.

Examples

# Use rstanarm package to quickly fit a model and get both a log-likelihood
# matrix and draws from the posterior predictive distribution
library("rstanarm")

# data from help("lm")
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
d <- data.frame(
  weight = c(ctl, trt),
  group = gl(2, 10, 20, labels = c("Ctl","Trt"))
)
fit <- stan_glm(weight ~ group, data = d)

#> trying deprecated constructor; please alert package maintainer

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 2.5e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.25 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.045101 seconds (Warm-up)
#>                0.046464 seconds (Sampling)
#>                0.091565 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 3.1e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.31 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.044581 seconds (Warm-up)
#>                0.044181 seconds (Sampling)
#>                0.088762 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 2e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.048517 seconds (Warm-up)
#>                0.044183 seconds (Sampling)
#>                0.0927 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 1.1e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.11 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.044552 seconds (Warm-up)
#>                0.044468 seconds (Sampling)
#>                0.08902 seconds (Total)
#> 

yrep <- posterior_predict(fit)
dim(yrep)

#> [1] 4000   20

ll <- log_lik(fit)
lw <- psislw(-ll, cores = 2)$lw_smooth

#> Warning: Some Pareto k diagnostic values are slightly high. See help('pareto-k-diagnostic') for details.

dim(lw)

#> [1] 4000   20

E_loo(yrep, lw, type = "mean")

#>  [1] 5.138671 4.985160 5.016090 4.915210 5.080892 5.077818 4.994131 5.075250
#>  [9] 4.992551 5.001659 4.644745 4.713694 4.696196 4.773859 4.534438 4.748797
#> [17] 4.523078 4.623510 4.711757 4.665274

E_loo(yrep, lw, type = "var")

#>  [1] 0.5946351 0.6011310 0.6076841 0.5489710 0.6124135 0.6378487 0.6234596
#>  [8] 0.6179831 0.6181676 0.5989046 0.6289565 0.6127202 0.6286769 0.5529180
#> [15] 0.5205391 0.5717348 0.4846393 0.6264684 0.6215462 0.6415826

E_loo(yrep, lw, type = "quantile", probs = 0.5) # median

#>  [1] 5.138123 4.989043 5.023238 4.930933 5.076973 5.072379 4.996906 5.079048
#>  [9] 4.981956 4.985157 4.629981 4.715855 4.698259 4.772829 4.525018 4.749758
#> [17] 4.531188 4.631676 4.710634 4.661785

E_loo(yrep, lw, type = "quantile", probs = c(0.1, 0.9))

#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 4.168906 4.022142 4.018365 3.964330 4.078852 4.084518 4.001782 4.083763
#> [2,] 6.113732 5.937602 5.987116 5.831365 6.067328 6.072594 5.965669 6.042356
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 4.017478 4.038297 3.666715 3.741089 3.695300 3.835308 3.616535 3.814350
#> [2,] 5.991321 5.974115 5.651807 5.702884 5.670938 5.698503 5.425953 5.706992
#>         [,17]    [,18]    [,19]    [,20]
#> [1,] 3.644202 3.609945 3.720679 3.649208
#> [2,] 5.398620 5.608497 5.704476 5.698210