Compute weighted expectations

The E_loo function computes weighted expectations (means, variances, quantiles) using the importance weights obtained from the PSIS smoothing procedure. The expectations estimated by the E_loo function assume that the PSIS approximation is working well. A small Pareto k estimate is necessary, but not sufficient, for E_loo to give reliable estimates. Additional diagnostic checks for gauging the reliability of the estimates are in development and will be added in a future release.

E_loo(x, psis_object, ...)

# S3 method for default
E_loo(x, psis_object, ..., type = c("mean", "variance",
  "quantile"), probs = NULL, log_ratios = NULL)

# S3 method for matrix
E_loo(x, psis_object, ..., type = c("mean", "variance",
  "quantile"), probs = NULL, log_ratios = NULL)

Arguments

x
A numeric vector or matrix.

psis_object
An object returned by psis.

...
Arguments passed to individual methods.

type
The type of expectation to compute. The options are "mean", "variance", and "quantile".

probs
For computing quantiles, a vector of probabilities.

log_ratios
Optionally, a vector or matrix (the same dimensions as x) of raw (not smoothed) log ratios. If working with log-likelihood values, the log ratios are the negative of those values. If log_ratios is specified we are able to compute Pareto k diagnostics specific to E_loo.

Value

A named list with the following components:

value

The result of the computation. For the matrix method, value is a vector with ncol(x) elements, with one exception: when type is "quantile" and multiple values are specified in probs the value component of the returned object is a length(probs) by ncol(x) matrix. For the default/vector method the value component is scalar, with one exception: when type is "quantile" and multiple values are specified in probs the value component is a vector with length(probs) elements.

pareto_k

Function-specific diagnostic. If log_ratios is not specified when calling E_loo, pareto_k will be NULL. Otherwise, for the matrix method it will be a vector of length ncol(x) containing estimates of the shape parameter \(k\) of the generalized Pareto distribution. For the default/vector method, the estimate is a scalar.

Examples

# Use rstanarm package to quickly fit a model and get both a log-likelihood
# matrix and draws from the posterior predictive distribution
library("rstanarm")

# data from help("lm")
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
d <- data.frame(
  weight = c(ctl, trt),
  group = gl(2, 10, 20, labels = c("Ctl","Trt"))
)
fit <- stan_glm(weight ~ group, data = d)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 0.00015 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.067668 seconds (Warm-up)
#>                0.0619 seconds (Sampling)
#>                0.129568 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 2.3e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.23 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.063336 seconds (Warm-up)
#>                0.069948 seconds (Sampling)
#>                0.133284 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 3.3e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.33 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.071757 seconds (Warm-up)
#>                0.067747 seconds (Sampling)
#>                0.139504 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 3.9e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.39 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.058925 seconds (Warm-up)
#>                0.056239 seconds (Sampling)
#>                0.115164 seconds (Total)
#> 

yrep <- posterior_predict(fit)
dim(yrep)

#> [1] 4000   20

log_ratios <- -1 * log_lik(fit)
dim(log_ratios)

#> [1] 4000   20

r_eff <- relative_eff(exp(-log_ratios), chain_id = rep(1:4, each = 1000))
psis_object <- psis(log_ratios, r_eff = r_eff, cores = 2)

#> Warning: Some Pareto k diagnostic values are slightly high. See help('pareto-k-diagnostic') for details.

E_loo(yrep, psis_object, type = "mean")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.128897 4.984826 5.011531 4.908784 5.091790 5.073746 5.005896 5.082598
#>  [9] 4.990450 4.998032 4.643375 4.721066 4.703745 4.774520 4.509724 4.754599
#> [17] 4.505512 4.635344 4.709592 4.667202
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "var")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 0.0001754501 0.0001814906 0.0003149217 0.0002419038 0.0002111947
#>  [6] 0.0002235507 0.0003156031 0.0002107437 0.0002862320 0.0003145455
#> [11] 0.0002762232 0.0002297163 0.0002857804 0.0002434963 0.0004458073
#> [16] 0.0002065586 0.0004816739 0.0002698437 0.0002521878 0.0002962633
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = 0.5) # median

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.147329 4.989900 5.022945 4.924012 5.082351 5.068881 5.008923 5.078301
#>  [9] 4.999786 5.005507 4.632652 4.714507 4.709175 4.792277 4.495354 4.747111
#> [17] 4.513092 4.635710 4.710216 4.674140
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = c(0.1, 0.9))

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 4.165266 4.005794 4.005688 3.930393 4.098168 4.091638 4.021810 4.104155
#> [2,] 6.051758 5.959775 5.973330 5.871928 6.089724 6.083237 5.994043 6.072166
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 3.991309 4.004000 3.660305 3.733439 3.665925 3.835678 3.583881 3.804996
#> [2,] 5.996076 5.983455 5.652360 5.753795 5.717199 5.689535 5.440069 5.721900
#>         [,17]    [,18]    [,19]    [,20]
#> [1,] 3.639736 3.645159 3.760078 3.644717
#> [2,] 5.365498 5.621357 5.691080 5.672981
#> 
#> $pareto_k
#> NULL
#> 

# To get Pareto k diagnostic with E_loo we also need to provide the negative
# log-likelihood values using the log_ratios argument.
E_loo(yrep, psis_object, type = "mean", log_ratios = log_ratios)

#> $value
#>  [1] 5.128897 4.984826 5.011531 4.908784 5.091790 5.073746 5.005896 5.082598
#>  [9] 4.990450 4.998032 4.643375 4.721066 4.703745 4.774520 4.509724 4.754599
#> [17] 4.505512 4.635344 4.709592 4.667202
#> 
#> $pareto_k
#>  [1]  0.200433088  0.176415666  0.017306495  0.347158374  0.241378021
#>  [6]  0.114432427  0.031429355  0.264012302  0.022469390  0.104256912
#> [11]  0.145207833  0.158421630  0.030003448  0.508770508  0.558038799
#> [16]  0.385210961  0.616980366 -0.003966085  0.068256269  0.085821650
#>