Compute weighted expectations

The E_loo function computes weighted expectations (means, variances, quantiles) using the importance weights obtained from the PSIS smoothing procedure. The expectations estimated by the E_loo function assume that the PSIS approximation is working well. A small Pareto k estimate is necessary, but not sufficient, for E_loo to give reliable estimates. Additional diagnostic checks for gauging the reliability of the estimates are in development and will be added in a future release.

E_loo(x, psis_object, ...)

# S3 method for default
E_loo(x, psis_object, ..., type = c("mean", "variance",
  "quantile"), probs = NULL, log_ratios = NULL)

# S3 method for matrix
E_loo(x, psis_object, ..., type = c("mean", "variance",
  "quantile"), probs = NULL, log_ratios = NULL)

Arguments

x
A numeric vector or matrix.

psis_object
An object returned by psis.

...
Arguments passed to individual methods.

type
The type of expectation to compute. The options are "mean", "variance", and "quantile".

probs
For computing quantiles, a vector of probabilities.

log_ratios
Optionally, a vector or matrix (the same dimensions as x) of raw (not smoothed) log ratios. If working with log-likelihood values, the log ratios are the negative of those values. If log_ratios is specified we are able to compute Pareto k diagnostics specific to E_loo.

Value

A named list with the following components:

value

The result of the computation. For the matrix method, value is a vector with ncol(x) elements, with one exception: when type is "quantile" and multiple values are specified in probs the value component of the retured object is a length(probs) by ncol(x) matrix. For the default/vector method the value component is scalar, with one exception: when type is "quantile" and multiple values are specified in probs the value component is a vector with length(probs) elements.

pareto_k

Function-specific diagnostic. If log_ratios is not specified when calling E_loo, pareto_k will be NULL. Otherwise, for the matrix method it will be a vector of length ncol(x) containing estimates of the shape parameter \(k\) of the generalized Pareto distribution. For the default/vector method, the estimate is a scalar.

Examples

# Use rstanarm package to quickly fit a model and get both a log-likelihood
# matrix and draws from the posterior predictive distribution
library("rstanarm")

#> Loading required package: Rcpp

#> rstanarm (Version 2.17.4, packaged: )

#> - Do not expect the default priors to remain the same in future rstanarm versions.

#> Thus, R scripts should specify priors explicitly, even if they are just the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#> options(mc.cores = parallel::detectCores())

#> - Plotting theme set to bayesplot::theme_default().

# data from help("lm")
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
d <- data.frame(
  weight = c(ctl, trt),
  group = gl(2, 10, 20, labels = c("Ctl","Trt"))
)
fit <- stan_glm(weight ~ group, data = d)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 0.000107 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 1.07 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.042826 seconds (Warm-up)
#>                0.039434 seconds (Sampling)
#>                0.08226 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 1.2e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.12 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.040688 seconds (Warm-up)
#>                0.042527 seconds (Sampling)
#>                0.083215 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 4.7e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.47 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.052404 seconds (Warm-up)
#>                0.043286 seconds (Sampling)
#>                0.09569 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 1.3e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.043984 seconds (Warm-up)
#>                0.044409 seconds (Sampling)
#>                0.088393 seconds (Total)
#> 

yrep <- posterior_predict(fit)
dim(yrep)

#> [1] 4000   20

ll <- log_lik(fit)
dim(ll)

#> [1] 4000   20

r_eff <- relative_eff(exp(-ll), chain_id = rep(1:4, each = 1000))
psis_object <- psis(log_ratios = -ll, r_eff = r_eff, cores = 2)

#> Warning: Some Pareto k diagnostic values are slightly high. See help('pareto-k-diagnostic') for details.

E_loo(yrep, psis_object, type = "mean")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.128833 4.984844 5.011531 4.908768 5.091755 5.073753 5.005882 5.082594
#>  [9] 4.990460 4.998012 4.643382 4.721145 4.703733 4.774520 4.509268 4.754449
#> [17] 4.505448 4.635494 4.709564 4.667213
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "var")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 0.0001886264 0.0002327015 0.0003506593 0.0002609670 0.0002592614
#>  [6] 0.0002670433 0.0003511051 0.0002536173 0.0003262130 0.0003494336
#> [11] 0.0003217072 0.0002779428 0.0003213056 0.0002447689 0.0003370669
#> [16] 0.0002385899 0.0003215824 0.0003205583 0.0002912583 0.0003326237
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = 0.5) # median

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.147247 4.989908 5.022941 4.924001 5.082348 5.068875 5.008909 5.078303
#>  [9] 4.999787 5.005506 4.632665 4.714657 4.709132 4.792277 4.495063 4.747107
#> [17] 4.513051 4.635817 4.710215 4.674142
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = c(0.1, 0.9))

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>          [,1]     [,2]     [,3]     [,4]     [,5]    [,6]     [,7]     [,8]
#> [1,] 4.165263 4.005839 4.005688 3.930421 4.098159 4.09161 4.021794 4.104154
#> [2,] 6.051625 5.959816 5.973289 5.871907 6.089659 6.08324 5.994003 6.072161
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 3.991392 4.003963 3.660310 3.733489 3.665923 3.835678 3.580245 3.804866
#> [2,] 5.996076 5.983452 5.652362 5.753682 5.717218 5.689535 5.439665 5.719812
#>         [,17]    [,18]    [,19]    [,20]
#> [1,] 3.639597 3.645458 3.760062 3.644717
#> [2,] 5.367175 5.621363 5.691079 5.672983
#> 
#> $pareto_k
#> NULL
#> 

# To get Pareto k diagnostic with E_loo we also need to provide the negative
# log-likelihood values using the log_ratios argument.
E_loo(yrep, psis_object, type = "mean", log_ratios = -ll)

#> $value
#>  [1] 5.128833 4.984844 5.011531 4.908768 5.091755 5.073753 5.005882 5.082594
#>  [9] 4.990460 4.998012 4.643382 4.721145 4.703733 4.774520 4.509268 4.754449
#> [17] 4.505448 4.635494 4.709564 4.667213
#> 
#> $pareto_k
#>  [1] 0.180522216 0.172368616 0.016644363 0.337234123 0.234526375 0.112964343
#>  [7] 0.049724409 0.271361198 0.006507165 0.072583719 0.159187834 0.192203023
#> [13] 0.012472731 0.508770508 0.496375655 0.384671512 0.588086034 0.014594418
#> [19] 0.075905565 0.082729696
#>