Compute weighted expectations

The E_loo() function computes weighted expectations (means, variances, quantiles) using the importance weights obtained from the PSIS smoothing procedure. The expectations estimated by the E_loo() function assume that the PSIS approximation is working well. A small Pareto k estimate is necessary, but not sufficient, for E_loo() to give reliable estimates. Additional diagnostic checks for gauging the reliability of the estimates are in development and will be added in a future release.

E_loo(x, psis_object, ...)

# S3 method for default
E_loo(
  x,
  psis_object,
  ...,
  type = c("mean", "variance", "quantile"),
  probs = NULL,
  log_ratios = NULL
)

# S3 method for matrix
E_loo(
  x,
  psis_object,
  ...,
  type = c("mean", "variance", "quantile"),
  probs = NULL,
  log_ratios = NULL
)

Arguments

x
A numeric vector or matrix.

psis_object
An object returned by psis().

...
Arguments passed to individual methods.

type
The type of expectation to compute. The options are "mean", "variance", and "quantile".

probs
For computing quantiles, a vector of probabilities.

log_ratios
Optionally, a vector or matrix (the same dimensions as x) of raw (not smoothed) log ratios. If working with log-likelihood values, the log ratios are the negative of those values. If log_ratios is specified we are able to compute Pareto k diagnostics specific to E_loo().

Value

A named list with the following components:

value

The result of the computation.For the matrix method, value is a vector with ncol(x) elements, with one exception: when type="quantile" and multiple values are specified in probs the value component of the returned object is a length(probs) by ncol(x) matrix.For the default/vector method the value component is scalar, with one exception: when type is "quantile" and multiple values are specified in probs the value component is a vector with length(probs) elements.

pareto_k

Function-specific diagnostic.If log_ratios is not specified when calling E_loo(), pareto_k will be NULL. Otherwise, for the matrix method it will be a vector of length ncol(x) containing estimates of the shape parameter \(k\) of the generalized Pareto distribution. For the default/vector method, the estimate is a scalar.

Examples

# \donttest{
# Use rstanarm package to quickly fit a model and get both a log-likelihood
# matrix and draws from the posterior predictive distribution
library("rstanarm")

#> Loading required package: Rcpp

#> This is rstanarm version 2.19.3

#> - See https://mc-stan.org/rstanarm/articles/priors for changes to default priors!

#> - Default priors may change, so it's safest to specify priors, even if equivalent to the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#>   options(mc.cores = parallel::detectCores())

# data from help("lm")
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
d <- data.frame(
  weight = c(ctl, trt),
  group = gl(2, 10, 20, labels = c("Ctl","Trt"))
)
fit <- stan_glm(weight ~ group, data = d, refresh = 0)
yrep <- posterior_predict(fit)
dim(yrep)

#> [1] 4000   20

log_ratios <- -1 * log_lik(fit)
dim(log_ratios)

#> [1] 4000   20

r_eff <- relative_eff(exp(-log_ratios), chain_id = rep(1:4, each = 1000))
psis_object <- psis(log_ratios, r_eff = r_eff, cores = 2)

E_loo(yrep, psis_object, type = "mean")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.130161 4.942460 5.023006 4.889126 5.079624 5.057261 4.993542 5.100044
#>  [9] 4.991074 5.000810 4.634745 4.742161 4.694916 4.797833 4.541844 4.734143
#> [17] 4.511471 4.659041 4.715469 4.663138
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "var")

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 0.0001817459 0.0002085514 0.0002874214 0.0002135392 0.0002086019
#>  [6] 0.0002452564 0.0003097210 0.0002085195 0.0002753723 0.0003525226
#> [11] 0.0002709946 0.0002145444 0.0002599368 0.0002203694 0.0003086298
#> [16] 0.0001806339 0.0004536805 0.0002728438 0.0002478241 0.0002778719
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = 0.5) # median

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>  [1] 5.137362 4.946344 5.030535 4.892565 5.076470 5.063158 4.987480 5.101923
#>  [9] 4.994185 5.025377 4.625960 4.739336 4.706874 4.791667 4.563775 4.741163
#> [17] 4.522306 4.653584 4.720618 4.652127
#> 
#> $pareto_k
#> NULL
#> 

E_loo(yrep, psis_object, type = "quantile", probs = c(0.1, 0.9))

#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.

#> $value
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 4.191633 3.914375 4.039848 3.942529 4.072185 4.054182 3.985790 4.100668
#> [2,] 6.065339 5.954295 5.994522 5.805594 6.074038 6.050066 5.982853 6.071363
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 3.974670 3.947726 3.638488 3.773463 3.704672 3.872623 3.615117 3.776612
#> [2,] 6.004577 6.033418 5.644392 5.738865 5.682099 5.734365 5.455974 5.690030
#>        [,17]    [,18]    [,19]    [,20]
#> [1,] 3.66190 3.689718 3.724255 3.687103
#> [2,] 5.38953 5.650770 5.745351 5.679150
#> 
#> $pareto_k
#> NULL
#> 

# To get Pareto k diagnostic with E_loo we also need to provide the negative
# log-likelihood values using the log_ratios argument.
E_loo(yrep, psis_object, type = "mean", log_ratios = log_ratios)

#> $value
#>  [1] 5.130161 4.942460 5.023006 4.889126 5.079624 5.057261 4.993542 5.100044
#>  [9] 4.991074 5.000810 4.634745 4.742161 4.694916 4.797833 4.541844 4.734143
#> [17] 4.511471 4.659041 4.715469 4.663138
#> 
#> $pareto_k
#>  [1]  0.17330278  0.21427164  0.16604684  0.27986236  0.10064752  0.09979168
#>  [7]  0.17187747  0.23472381  0.21594104  0.15597476  0.10347666  0.12778891
#> [13]  0.05270634  0.21715028  0.41637775  0.16749605  0.47254928  0.21880378
#> [19]  0.06410958 -0.01645188
#> 

# }