Compute weighted expectations

The E_loo function computes weighted expectations (means, variances, quantiles) using the smoothed importance weights obtained from the PSIS procedure. The expectations estimated by the E_loo function assume that the PSIS approximation is working well. A small Pareto k estimate is necessary, but not sufficient, for E_loo to give reliable estimates. Additional diagnostic checks for gauging the reliability of the estimates are in development and will be added in a future release.

E_loo(x, lw, ...)

# S3 method for default
E_loo(x, lw, ..., type = c("mean", "var", "quantile"),
  probs)

# S3 method for matrix
E_loo(x, lw, ..., type = c("mean", "var", "quantile"), probs)

Arguments

x
A numeric matrix or vector.

lw
A numeric matrix (or vector) of smoothed log-weights with the same dimensions (or length) as x. Typically this will be the "lw_smooth" element of the list returned by psislw.

...
For the generic function, arguments passed to the individual methods.

type
The type of expectation to compute. The options are "mean", "var" (variance), and "quantile".

probs
A vector of probabilities. Ignored unless type is "quantile".

Value

The matrix method returns a vector with ncol(x) elements, with one exception: when type is "quantile" and multiple values are specified in probs the returned object is a length(probs) by ncol(x) matrix.

The default/vector method returns a scalar, with one exception: when type is "quantile" and multiple values are specified in probs the returned object is a vector with length(probs) elements.

Examples

# Use rstanarm package to quickly fit a model and get both a log-likelihood
# matrix and draws from the posterior predictive distribution
library("rstanarm")

# data from help("lm")
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
d <- data.frame(
  weight = c(ctl, trt),
  group = gl(2, 10, 20, labels = c("Ctl","Trt"))
)
fit <- stan_glm(weight ~ group, data = d)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Chain 1, Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1, Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1, Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1, Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1, Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1, Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1, Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1, Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1, Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1, Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1, Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1, Iteration: 2000 / 2000 [100%]  (Sampling)
#>  Elapsed Time: 0.034536 seconds (Warm-up)
#>                0.041687 seconds (Sampling)
#>                0.076223 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Chain 2, Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2, Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2, Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2, Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2, Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2, Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2, Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2, Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2, Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2, Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2, Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2, Iteration: 2000 / 2000 [100%]  (Sampling)
#>  Elapsed Time: 0.036694 seconds (Warm-up)
#>                0.038533 seconds (Sampling)
#>                0.075227 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Chain 3, Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3, Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3, Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3, Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3, Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3, Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3, Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3, Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3, Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3, Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3, Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3, Iteration: 2000 / 2000 [100%]  (Sampling)
#>  Elapsed Time: 0.03942 seconds (Warm-up)
#>                0.040074 seconds (Sampling)
#>                0.079494 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Chain 4, Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4, Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4, Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4, Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4, Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4, Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4, Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4, Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4, Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4, Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4, Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4, Iteration: 2000 / 2000 [100%]  (Sampling)
#>  Elapsed Time: 0.047375 seconds (Warm-up)
#>                0.040055 seconds (Sampling)
#>                0.08743 seconds (Total)
#> 

yrep <- posterior_predict(fit)
dim(yrep)

#> [1] 4000   20

ll <- log_lik(fit)
lw <- psislw(-ll, cores = 2)$lw_smooth
dim(lw)

#> [1] 4000   20

E_loo(yrep, lw, type = "mean")

#>  [1] 5.138855 4.997955 5.017800 4.926316 5.102033 5.073566 5.007486 5.092255
#>  [9] 5.001263 5.003172 4.636040 4.710531 4.684775 4.767932 4.525041 4.741099
#> [17] 4.522131 4.615086 4.701964 4.663462

E_loo(yrep, lw, type = "var")

#>  [1] 0.6180670 0.6401770 0.6516194 0.5844079 0.6692286 0.6737588 0.6838490
#>  [8] 0.6780532 0.6707002 0.6658463 0.6741706 0.6778545 0.6720629 0.6009230
#> [15] 0.5711960 0.6306080 0.5438892 0.6642430 0.6648305 0.6739408

E_loo(yrep, lw, type = "quantile", probs = 0.5) # median

#>  [1] 5.136836 4.999639 5.026892 4.950176 5.093168 5.059823 4.999824 5.077163
#>  [9] 4.986227 5.005911 4.628662 4.705696 4.687115 4.771179 4.528629 4.742519
#> [17] 4.554246 4.623864 4.707404 4.658413

E_loo(yrep, lw, type = "quantile", probs = c(0.1, 0.9))

#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 4.156855 3.986567 3.977106 3.932966 4.081145 4.052193 4.001953 4.072532
#> [2,] 6.127149 6.011426 6.030150 5.847882 6.124638 6.113332 6.044192 6.118164
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 4.012326 3.985163 3.613652 3.702373 3.646804 3.772815 3.545232 3.753980
#> [2,] 6.021163 6.018761 5.667944 5.738859 5.709658 5.770713 5.462676 5.707672
#>         [,17]    [,18]    [,19]    [,20]
#> [1,] 3.534326 3.589634 3.674429 3.626977
#> [2,] 5.431275 5.627683 5.713758 5.683448