Compute weighted expectations

The E_loo() function computes weighted expectations (means, variances, quantiles) using the importance weights obtained from the PSIS smoothing procedure. The expectations estimated by the E_loo() function assume that the PSIS approximation is working well. A small Pareto k estimate is necessary, but not sufficient, for E_loo() to give reliable estimates. Additional diagnostic checks for gauging the reliability of the estimates are in development and will be added in a future release.

E_loo(x, psis_object, ...)

# S3 method for default
E_loo(
  x,
  psis_object,
  ...,
  type = c("mean", "variance", "quantile"),
  probs = NULL,
  log_ratios = NULL
)

# S3 method for matrix
E_loo(
  x,
  psis_object,
  ...,
  type = c("mean", "variance", "quantile"),
  probs = NULL,
  log_ratios = NULL
)

Arguments

x

A numeric vector or matrix.

psis_object

An object returned by psis().

...

Arguments passed to individual methods.

type

The type of expectation to compute. The options are "mean", "variance", and "quantile".

probs

For computing quantiles, a vector of probabilities.

log_ratios

Optionally, a vector or matrix (the same dimensions as x) of raw (not smoothed) log ratios. If working with log-likelihood values, the log ratios are the negative of those values. If log_ratios is specified we are able to compute Pareto k diagnostics specific to E_loo().

Value

A named list with the following components:

value

The result of the computation.

For the matrix method, value is a vector with ncol(x) elements, with one exception: when type="quantile" and multiple values are specified in probs the value component of the returned object is a length(probs) by ncol(x) matrix.

For the default/vector method the value component is scalar, with one exception: when type is "quantile" and multiple values are specified in probs the value component is a vector with length(probs) elements.

pareto_k

Function-specific diagnostic.

If log_ratios is not specified when calling E_loo(), pareto_k will be NULL. Otherwise, for the matrix method it will be a vector of length ncol(x) containing estimates of the shape parameter \(k\) of the generalized Pareto distribution. For the default/vector method, the estimate is a scalar.

Examples

# \donttest{
# Use rstanarm package to quickly fit a model and get both a log-likelihood
# matrix and draws from the posterior predictive distribution
library("rstanarm")
#> Loading required package: Rcpp
#> This is rstanarm version 2.21.3
#> - See https://mc-stan.org/rstanarm/articles/priors for changes to default priors!
#> - Default priors may change, so it's safest to specify priors, even if equivalent to the defaults.
#> - For execution on a local, multicore CPU with excess RAM we recommend calling
#>   options(mc.cores = parallel::detectCores())

# data from help("lm")
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
d <- data.frame(
  weight = c(ctl, trt),
  group = gl(2, 10, 20, labels = c("Ctl","Trt"))
)
fit <- stan_glm(weight ~ group, data = d, refresh = 0)
yrep <- posterior_predict(fit)
dim(yrep)
#> [1] 4000   20

log_ratios <- -1 * log_lik(fit)
dim(log_ratios)
#> [1] 4000   20

r_eff <- relative_eff(exp(-log_ratios), chain_id = rep(1:4, each = 1000))
psis_object <- psis(log_ratios, r_eff = r_eff, cores = 2)

E_loo(yrep, psis_object, type = "mean")
#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.
#> $value
#>  [1] 5.159049 4.966521 5.007666 4.936129 5.110558 5.108392 5.030244 5.080837
#>  [9] 5.000796 5.029638 4.635559 4.701159 4.695538 4.768058 4.542233 4.762575
#> [17] 4.509362 4.621553 4.713217 4.669388
#> 
#> $pareto_k
#> NULL
#> 
E_loo(yrep, psis_object, type = "var")
#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.
#> $value
#>  [1] 0.0002037422 0.0002121794 0.0003120603 0.0002381440 0.0001996315
#>  [6] 0.0002234251 0.0003140785 0.0001937419 0.0002647202 0.0003179773
#> [11] 0.0002882070 0.0002318150 0.0002846632 0.0002192418 0.0003495323
#> [16] 0.0002294693 0.0005414286 0.0002557699 0.0002598545 0.0003085766
#> 
#> $pareto_k
#> NULL
#> 
E_loo(yrep, psis_object, type = "quantile", probs = 0.5) # median
#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.
#> $value
#>  [1] 5.151455 4.969630 5.006205 4.911495 5.085145 5.096823 5.021451 5.083281
#>  [9] 4.989998 5.028857 4.635323 4.705774 4.695495 4.778008 4.540621 4.760443
#> [17] 4.517806 4.628075 4.705734 4.659420
#> 
#> $pareto_k
#> NULL
#> 
E_loo(yrep, psis_object, type = "quantile", probs = c(0.1, 0.9))
#> Warning: 'log_ratios' not specified. Can't compute k-hat diagnostic.
#> $value
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 4.195856 3.929749 3.985757 4.023786 4.141688 4.118301 4.012345 4.104410
#> [2,] 6.130801 5.963534 6.014129 5.864966 6.115942 6.114645 6.018165 6.041274
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 3.994901 4.014882 3.641831 3.713065 3.696238 3.854874 3.629619 3.789748
#> [2,] 5.996236 6.054583 5.625478 5.673327 5.711968 5.690776 5.431133 5.748500
#>         [,17]    [,18]    [,19]    [,20]
#> [1,] 3.627065 3.646186 3.724462 3.647525
#> [2,] 5.434909 5.618534 5.703162 5.671337
#> 
#> $pareto_k
#> NULL
#> 

# To get Pareto k diagnostic with E_loo we also need to provide the negative
# log-likelihood values using the log_ratios argument.
E_loo(yrep, psis_object, type = "mean", log_ratios = log_ratios)
#> $value
#>  [1] 5.159049 4.966521 5.007666 4.936129 5.110558 5.108392 5.030244 5.080837
#>  [9] 5.000796 5.029638 4.635559 4.701159 4.695538 4.768058 4.542233 4.762575
#> [17] 4.509362 4.621553 4.713217 4.669388
#> 
#> $pareto_k
#>  [1]  0.34040926  0.14872826  0.05050987  0.30727724  0.22794724  0.15023183
#>  [7]  0.09826373  0.19426301  0.15156339  0.15956006  0.02110562  0.04121434
#> [13] -0.00999068  0.10847330  0.30014603  0.13221484  0.38781537  0.09850812
#> [19]  0.09001664  0.01124195
#> 
# }