Continuously ranked probability score

The crps() and scrps() functions and their loo_*() counterparts can be used to compute the continuously ranked probability score (CRPS) and scaled CRPS (SCRPS) (as defined by Bolin and Wallin, 2023). CRPS is a proper scoring rule, and strictly proper when the first moment of the predictive distribution is finite. Both can be expressed in terms of samples form the predictive distribution. See, for example, a paper by Gneiting and Raftery (2007) for a comprehensive discussion on CRPS.

Usage

crps(x, ...)

scrps(x, ...)

loo_crps(x, ...)

loo_scrps(x, ...)

# S3 method for class 'matrix'
crps(x, x2, y, ..., permutations = 1)

# S3 method for class 'numeric'
crps(x, x2, y, ..., permutations = 1)

# S3 method for class 'matrix'
loo_crps(
  x,
  x2,
  y,
  log_lik,
  ...,
  permutations = 1,
  r_eff = 1,
  cores = getOption("mc.cores", 1)
)

# S3 method for class 'matrix'
scrps(x, x2, y, ..., permutations = 1)

# S3 method for class 'numeric'
scrps(x, x2, y, ..., permutations = 1)

# S3 method for class 'matrix'
loo_scrps(
  x,
  x2,
  y,
  log_lik,
  ...,
  permutations = 1,
  r_eff = 1,
  cores = getOption("mc.cores", 1)
)

Arguments

x

A S by N matrix (draws by observations), or a vector of length S when only single observation is provided in y.

...

Passed on to E_loo() in the loo_*() version of these functions.

x2

Independent draws from the same distribution as draws in x. Should be of the identical dimension.

y

A vector of observations or a single value.

permutations

An integer, with default value of 1, specifying how many times the expected value of |X - X'| (|x - x2|) is computed. The row order of x2 is shuffled as elements x and x2 are typically drawn given the same values of parameters. This happens, e.g., when one calls posterior_predict() twice for a fitted rstanarm or brms model. Generating more permutations is expected to decrease the variance of the computed expected value.

log_lik

A log-likelihood matrix the same size as x.

r_eff

An optional vector of relative effective sample size estimates containing one element per observation. See psis() for details.

cores

The number of cores to use for parallelization of [psis()]. See psis() for details.

Value

A list containing two elements: estimates and pointwise. The former reports estimator and standard error and latter the pointwise values. Following Bolin & Wallin (2023), a larger value is better.

Details

To compute (S)CRPS, the user needs to provide two sets of draws, x and x2, from the predictive distribution. This is due to the fact that formulas used to compute CRPS involve an expectation of the absolute difference of x and x2, both having the same distribution. See the permutations argument, as well as Gneiting and Raftery (2007) for details.

References

Bolin, D., & Wallin, J. (2023). Local scale invariance and robustness of proper scoring rules. Statistical Science, 38(1):140-159.

Gneiting, T., & Raftery, A. E. (2007). Strictly Proper Scoring Rules, Prediction, and Estimation. Journal of the American Statistical Association, 102(477), 359–378.

Examples

# \dontrun{
# An example using rstanarm
library(rstanarm)
data("kidiq")
fit <- stan_glm(kid_score ~ mom_hs + mom_iq, data = kidiq)
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.3 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.032 seconds (Warm-up)
#> Chain 1:                0.062 seconds (Sampling)
#> Chain 1:                0.094 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 1.3e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.034 seconds (Warm-up)
#> Chain 2:                0.065 seconds (Sampling)
#> Chain 2:                0.099 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 1.2e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.12 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.038 seconds (Warm-up)
#> Chain 3:                0.064 seconds (Sampling)
#> Chain 3:                0.102 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 1.3e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.037 seconds (Warm-up)
#> Chain 4:                0.063 seconds (Sampling)
#> Chain 4:                0.1 seconds (Total)
#> Chain 4: 
ypred1 <- posterior_predict(fit)
ypred2 <- posterior_predict(fit)
crps(ypred1, ypred2, y = fit$y)
#> $estimates
#>    Estimate          SE 
#> -10.2201033   0.3462796 
#> 
#> $pointwise
#>          1          2          3          4          5          6          7 
#> -24.900751  -9.429146  -7.566819  -4.757940 -21.776399  -7.188831 -30.804403 
#>          8          9         10         11         12         13         14 
#>  -4.757819 -15.637096  -6.468492  -6.359427 -17.768061  -4.239904 -10.673001 
#>         15         16         17         18         19         20         21 
#> -11.561414  -6.692402  -5.390397  -4.318643  -4.445514 -12.452083  -5.665604 
#>         22         23         24         25         26         27         28 
#>  -7.378751  -7.754736  -4.345221 -11.053228  -4.832656  -8.728251  -5.861742 
#>         29         30         31         32         33         34         35 
#>  -4.313555  -4.057937 -15.823206 -30.627095  -6.826580 -12.101544  -5.089308 
#>         36         37         38         39         40         41         42 
#>  -9.514070 -13.078258  -4.691303  -5.944815 -18.215365  -9.966671  -5.378535 
#>         43         44         45         46         47         48         49 
#>  -8.406113  -4.406818 -20.135966  -5.749869 -24.510404  -4.460993 -17.431957 
#>         50         51         52         53         54         55         56 
#>  -4.765400  -6.663401  -5.582078 -19.759953 -14.754404  -5.325623  -8.904427 
#>         57         58         59         60         61         62         63 
#>  -6.197788  -4.937223  -5.284561  -5.616381  -9.268595  -8.761120  -4.327131 
#>         64         65         66         67         68         69         70 
#>  -5.930424 -23.483479 -25.138686 -13.046133  -4.409787  -5.123954  -5.424425 
#>         71         72         73         74         75         76         77 
#>  -5.830134  -5.192773  -4.421525 -14.751313 -11.591811 -16.786968  -4.808163 
#>         78         79         80         81         82         83         84 
#>  -4.873056 -20.363815  -4.659020 -11.978782 -16.056477  -8.690020 -10.145217 
#>         85         86         87         88         89         90         91 
#>  -6.371287  -4.215216 -34.773780  -5.828478  -4.379739  -7.326905  -6.964549 
#>         92         93         94         95         96         97         98 
#> -10.216889  -8.668180  -5.046801 -11.936571  -4.861803  -4.321718  -6.345546 
#>         99        100        101        102        103        104        105 
#> -13.690298 -16.683949  -4.866824 -15.809229 -10.871605  -4.586989  -8.808631 
#>        106        107        108        109        110        111        112 
#>  -4.481927 -12.945530  -7.816259  -5.494882  -8.391380 -29.463484 -13.190499 
#>        113        114        115        116        117        118        119 
#> -26.475558  -7.355715  -5.609233  -5.046325 -25.125862 -29.227117  -5.179128 
#>        120        121        122        123        124        125        126 
#> -15.897256 -11.517018  -5.355114  -4.743005  -4.542042 -19.157431 -13.700954 
#>        127        128        129        130        131        132        133 
#> -11.060062  -5.684123  -4.306502  -8.767889 -23.833348  -8.716688  -6.633172 
#>        134        135        136        137        138        139        140 
#>  -5.609098  -7.071051 -27.677424 -12.652698  -8.112550  -4.550719  -6.658283 
#>        141        142        143        144        145        146        147 
#>  -4.024724  -4.785121  -4.542477  -5.888854  -4.471665  -4.260616  -6.749968 
#>        148        149        150        151        152        153        154 
#> -23.320616 -15.247624 -12.348571  -5.178561 -28.495674  -4.154838  -9.125980 
#>        155        156        157        158        159        160        161 
#>  -8.651110  -4.695778  -7.205551  -7.762747 -10.042472  -4.724869  -5.419278 
#>        162        163        164        165        166        167        168 
#>  -7.321711  -5.710599  -5.684938  -5.157638  -8.308155  -8.532738  -4.083349 
#>        169        170        171        172        173        174        175 
#> -15.950781  -6.366487  -4.809156  -5.605162  -5.755546  -6.754661 -12.383820 
#>        176        177        178        179        180        181        182 
#>  -4.333933  -4.414126  -6.459915  -4.825279  -4.501981  -5.024166  -5.309689 
#>        183        184        185        186        187        188        189 
#>  -4.292834 -13.955145 -19.676055 -14.354093  -5.286629 -20.570589  -4.099824 
#>        190        191        192        193        194        195        196 
#> -10.327451  -8.586051  -4.618217  -7.028630  -6.801230  -7.308523  -5.014850 
#>        197        198        199        200        201        202        203 
#>  -6.764636 -11.187975  -6.635365  -4.580954  -4.725686  -5.049591  -7.791539 
#>        204        205        206        207        208        209        210 
#>  -4.811319  -8.697904  -7.435702  -7.453368  -4.270350 -15.245878  -4.791924 
#>        211        212        213        214        215        216        217 
#>  -7.842196 -12.926809 -39.772207 -19.405911 -11.434198  -6.739385  -6.149844 
#>        218        219        220        221        222        223        224 
#>  -7.846718  -5.139589  -6.247213  -7.814136 -22.926283  -5.032341  -5.200068 
#>        225        226        227        228        229        230        231 
#>  -5.528244  -5.002408  -4.802062  -6.562702 -15.494864  -9.674893  -4.705059 
#>        232        233        234        235        236        237        238 
#>  -4.274015  -8.972172  -7.139933  -6.944002  -4.425469  -7.772937  -4.217749 
#>        239        240        241        242        243        244        245 
#>  -4.504299  -6.113204  -4.403393 -21.564528  -4.113582 -13.595483  -6.663990 
#>        246        247        248        249        250        251        252 
#> -15.946724  -4.254775 -26.205819  -6.977809  -4.696165  -5.512515  -8.670659 
#>        253        254        255        256        257        258        259 
#> -10.980411  -4.761930  -4.379472  -5.854158 -10.457894 -10.903596  -8.402074 
#>        260        261        262        263        264        265        266 
#>  -4.369433  -4.794512  -6.834050  -6.691491 -15.167920  -6.513032  -6.754408 
#>        267        268        269        270        271        272        273 
#>  -6.094088 -21.070727 -19.046839  -8.256524  -4.761618 -27.479445 -43.050347 
#>        274        275        276        277        278        279        280 
#> -10.008444  -5.791209  -8.147274 -16.712370  -4.356783 -17.768102  -6.208954 
#>        281        282        283        284        285        286        287 
#> -13.991192 -19.411188 -21.483631  -4.081200  -4.213080 -43.131797 -18.057646 
#>        288        289        290        291        292        293        294 
#> -22.560500  -4.346636 -17.934602  -9.454650 -11.018643 -21.950537  -5.720397 
#>        295        296        297        298        299        300        301 
#>  -7.864779 -22.960938  -4.465751 -16.873918  -7.333854 -15.125827  -5.576326 
#>        302        303        304        305        306        307        308 
#> -23.463821  -4.656249  -5.861642  -4.114459  -4.525957 -37.019695  -5.745356 
#>        309        310        311        312        313        314        315 
#>  -4.251819 -28.584794 -12.486891 -29.241154 -10.702439  -4.357488  -4.326050 
#>        316        317        318        319        320        321        322 
#>  -4.240167  -5.449423  -6.822403 -10.114359  -6.980591  -4.855649  -4.462275 
#>        323        324        325        326        327        328        329 
#> -12.718060 -18.244126  -8.433017  -8.983453  -7.264701 -27.258033  -4.433268 
#>        330        331        332        333        334        335        336 
#>  -5.595724  -4.215245 -18.285949 -26.972903 -14.868302  -6.930501 -12.082793 
#>        337        338        339        340        341        342        343 
#> -14.953421 -19.599673  -8.106585  -5.695342 -21.703369  -4.586396 -12.748782 
#>        344        345        346        347        348        349        350 
#>  -3.955999 -16.971795 -16.534197 -29.610012 -16.826868  -5.513721  -6.937369 
#>        351        352        353        354        355        356        357 
#>  -8.680805 -10.166254 -11.578243  -4.163982 -19.910044 -27.248136 -15.752426 
#>        358        359        360        361        362        363        364 
#> -20.500953  -4.108723  -5.362101  -4.605677 -10.187313  -4.427260  -4.733038 
#>        365        366        367        368        369        370        371 
#> -11.077258  -4.469114 -12.863830 -30.380626 -11.802507  -4.564728  -8.907167 
#>        372        373        374        375        376        377        378 
#>  -4.336490  -8.981592  -4.676666 -15.353319 -15.790499 -18.287881  -4.549392 
#>        379        380        381        382        383        384        385 
#>  -4.367500  -6.379122 -13.979059  -9.809787  -7.731674  -9.709499 -11.793193 
#>        386        387        388        389        390        391        392 
#> -13.517956 -11.014441  -4.417755  -6.439535 -10.247766  -9.259853  -4.896456 
#>        393        394        395        396        397        398        399 
#> -19.425236 -20.799539  -4.224040  -5.224484 -20.871493 -24.302692  -4.486746 
#>        400        401        402        403        404        405        406 
#>  -6.671885  -4.364008  -4.453198  -5.858647  -4.548513 -13.924773  -5.371274 
#>        407        408        409        410        411        412        413 
#> -16.381598  -5.336469 -28.237185  -4.300187  -4.449397 -12.718340 -10.111967 
#>        414        415        416        417        418        419        420 
#>  -4.856456  -6.795591  -7.134434  -4.319025 -17.680057  -5.012004 -17.640512 
#>        421        422        423        424        425        426        427 
#> -18.777509  -4.649689 -11.628672  -9.772419 -23.594912  -4.159432  -4.310980 
#>        428        429        430        431        432        433        434 
#>  -7.447002 -16.413943 -12.962150  -6.031866 -19.480648  -4.140881  -8.171842 
#> 
loo_crps(ypred1, ypred2, y = fit$y, log_lik = log_lik(fit))
#> $estimates
#>    Estimate          SE 
#> -10.2912368   0.3503924 
#> 
#> $pointwise
#>   [1] -24.933879  -9.315822  -7.525864  -4.824768 -21.770014  -7.200150
#>   [7] -31.574259  -4.604895 -15.703184  -6.311157  -6.380353 -17.989547
#>  [13]  -4.384020 -10.906701 -11.895214  -6.778520  -5.298803  -4.366513
#>  [19]  -4.581238 -12.555362  -5.600648  -7.231979  -7.822592  -4.386619
#>  [25] -11.136209  -4.669964  -8.898885  -5.889211  -4.099102  -4.085196
#>  [31] -15.786696 -30.694010  -6.925735 -12.250704  -5.149778  -9.487870
#>  [37] -13.193748  -4.907658  -6.047509 -18.253374  -9.757294  -5.321013
#>  [43]  -8.461685  -4.204097 -20.371463  -5.800357 -24.832775  -4.680492
#>  [49] -17.616221  -4.826629  -6.577694  -5.741888 -19.953921 -14.610912
#>  [55]  -5.563438  -8.832792  -6.108729  -5.192572  -5.248929  -5.689373
#>  [61]  -9.280767  -8.754628  -4.482395  -6.218533 -23.439823 -25.276406
#>  [67] -13.174458  -4.401599  -5.057144  -5.316191  -6.113573  -5.123622
#>  [73]  -4.396391 -14.589720 -11.509078 -16.816923  -4.811600  -4.557785
#>  [79] -20.430259  -4.731993 -12.211152 -16.201319  -8.645641 -10.431037
#>  [85]  -6.336490  -4.397083 -35.183422  -5.486962  -4.541588  -7.545535
#>  [91]  -7.212858 -10.170751  -8.809499  -4.779650 -12.045204  -4.870838
#>  [97]  -4.331795  -6.379310 -13.908986 -16.632947  -4.843675 -15.821377
#> [103] -10.924184  -4.749359  -8.705180  -4.286338 -13.153582  -7.928730
#> [109]  -5.604011  -8.495927 -30.256012 -13.447452 -26.602579  -7.715142
#> [115]  -5.528883  -5.229427 -25.227062 -29.547532  -4.977191 -16.033259
#> [121] -11.326215  -5.313491  -4.821806  -4.602664 -19.029031 -13.657661
#> [127] -11.372596  -5.738648  -4.360213  -8.797475 -23.878702  -9.026219
#> [133]  -6.612040  -5.737628  -7.275742 -28.331954 -12.755155  -8.209566
#> [139]  -4.415978  -6.654696  -4.132792  -4.815997  -4.462456  -5.723029
#> [145]  -4.304174  -4.315007  -6.433260 -23.531288 -15.373328 -12.682956
#> [151]  -5.130176 -29.105098  -4.097046  -9.285432  -8.483337  -4.943988
#> [157]  -7.093225  -8.012079  -9.923160  -4.778062  -5.514009  -7.293528
#> [163]  -5.876239  -5.773723  -5.203442  -7.900729  -8.695076  -4.086002
#> [169] -16.143709  -6.599155  -4.756880  -5.532670  -5.658960  -6.728462
#> [175] -12.386419  -4.052497  -4.354663  -6.418926  -4.906346  -4.601980
#> [181]  -5.087123  -5.301824  -4.057640 -14.063282 -19.852121 -14.685886
#> [187]  -5.154350 -20.490822  -4.078968 -10.239380  -8.637872  -4.520613
#> [193]  -7.274317  -6.922198  -7.510263  -5.093955  -6.965876 -11.235202
#> [199]  -6.752734  -4.605235  -4.992255  -5.168896  -7.733479  -4.727830
#> [205]  -8.800838  -7.441018  -7.409364  -4.061908 -15.278618  -4.989973
#> [211]  -8.041577 -12.874809 -40.241786 -19.594878 -11.707529  -6.816399
#> [217]  -6.040175  -7.878443  -5.272143  -6.120493  -7.940211 -23.206908
#> [223]  -4.748237  -5.174912  -5.717732  -4.920493  -4.876465  -6.363641
#> [229] -15.580778  -9.744148  -4.804196  -4.092207  -9.259497  -6.918648
#> [235]  -6.806668  -4.532695  -7.664022  -4.098610  -4.515645  -5.821310
#> [241]  -4.496162 -21.903265  -4.444753 -13.556864  -6.719278 -16.157002
#> [247]  -4.287990 -26.189133  -7.022667  -4.608984  -5.633018  -8.483780
#> [253] -10.839214  -4.833024  -4.116561  -5.859998 -10.545403 -11.060097
#> [259]  -8.756308  -4.365141  -4.941151  -6.671308  -6.849694 -15.276376
#> [265]  -6.599560  -6.653796  -6.140402 -21.423890 -19.113553  -8.420758
#> [271]  -4.492846 -27.671840 -43.342498 -10.086732  -5.822858  -8.043752
#> [277] -16.814726  -4.228116 -17.873682  -6.322839 -14.289403 -19.688351
#> [283] -21.272080  -4.137823  -4.322546 -43.903977 -18.149632 -22.727622
#> [289]  -4.212608 -18.079781  -9.610608 -11.335897 -22.202077  -5.897228
#> [295]  -7.992434 -22.904991  -4.465663 -17.211939  -7.396938 -14.923470
#> [301]  -5.483246 -23.790123  -4.794689  -5.988188  -4.330138  -4.553125
#> [307] -36.766866  -5.604123  -4.188484 -29.472846 -12.409793 -29.303216
#> [313] -10.969151  -4.315479  -4.183555  -4.391536  -5.607787  -6.807094
#> [319] -10.132261  -6.816993  -4.945455  -4.487760 -12.619396 -18.507357
#> [325]  -8.676511  -8.887299  -6.953109 -27.276741  -4.685231  -5.700835
#> [331]  -4.332582 -18.381748 -27.144003 -15.255224  -6.948640 -12.113459
#> [337] -15.329050 -20.030661  -8.182007  -5.807114 -22.127739  -4.477266
#> [343] -12.749938  -4.136547 -17.083302 -16.841747 -30.049446 -16.863437
#> [349]  -5.400814  -7.049229  -8.680825 -10.236661 -11.613575  -4.194740
#> [355] -19.999945 -27.654102 -15.936786 -20.959039  -4.234658  -5.277492
#> [361]  -4.774292 -10.151355  -4.188110  -4.591837 -11.283158  -4.552224
#> [367] -12.658385 -30.710709 -11.970965  -4.548779  -9.026372  -4.412336
#> [373]  -8.825473  -4.762917 -15.566302 -15.855971 -18.457937  -4.430718
#> [379]  -4.420070  -6.582877 -14.237032  -9.916011  -7.935239  -9.787802
#> [385] -12.156401 -13.574812 -10.790766  -4.587365  -6.364079 -10.266527
#> [391]  -9.402814  -5.161094 -20.019991 -20.985641  -4.251236  -5.419401
#> [397] -20.918503 -24.488459  -4.237627  -6.728311  -4.457945  -4.215560
#> [403]  -6.092705  -4.508131 -14.323520  -5.485782 -16.304596  -5.231195
#> [409] -28.247476  -4.220712  -4.435665 -13.010244 -10.144669  -4.791444
#> [415]  -6.718298  -7.179668  -4.391743 -18.087046  -4.983353 -18.093200
#> [421] -19.014711  -4.654956 -11.937869  -9.917738 -24.019294  -4.102320
#> [427]  -4.393731  -7.575562 -16.615683 -12.918714  -5.810618 -19.929421
#> [433]  -4.356471  -8.064831
#> 
# }