Generalized Pareto distribution

Estimate the parameters \(k\) and \(\sigma\) of the generalized Pareto distribution, given a sample \(x\).

gpdfit(x)

Arguments

x
A numeric vector. The sample from which to estimate the parameters.

Value

A named list with components k and sigma.

Details

Here the parameter \(k\) is the negative of \(k\) in Zhang & Stephens (2009).

Note

This function is primarily intended for internal use, but is exported so that users can call it directly if interested in PSIS for purposes other than LOO-CV.

References

Zhang, J., and Stephens, M. A. (2009). A new and efficient estimation method for the generalized Pareto distribution. Technometrics 51, 316-325.

See also

psislw, pareto-k-diagnostic, loo-package