Named lists

Create a named list using specified names or, if names are omitted, using the names of the objects in the list. The code list(a = a, b = b) becomes nlist(a,b) and list(a = a, b = 2) becomes nlist(a, b = 2), etc.

nlist(...)

Arguments

...
Objects to include in the list.

Value

A named list.

Examples

# All variables already defined
a <- rnorm(100)
b <- mat.or.vec(10, 3)
nlist(a,b)

#> $a
#>   [1]  0.632021803 -0.691559915 -0.698551147 -2.126358177 -0.189012752
#>   [6]  0.205477737 -1.338316688 -1.155834491  0.185828351 -0.953053705
#>  [11]  1.035186689 -0.380511294 -0.804927356 -0.278846749  0.283314615
#>  [16]  1.065870044  0.284305792  1.499735240 -1.193944998 -0.299988252
#>  [21]  0.423484063 -0.475867734  2.043626802  0.008833826 -0.597469553
#>  [26]  0.893984469  0.145662593  0.159455138 -2.349083183  0.864750814
#>  [31] -0.044704628  1.108497705  0.365317198  0.067778239 -2.258293294
#>  [36] -0.307612797 -1.608490159 -1.172991402  1.116213939 -0.451311914
#>  [41]  1.139423503  0.147087735  0.223140303 -1.797794624 -0.257989149
#>  [46] -1.456310534  0.553609616 -1.424251381  1.181560514  0.070606527
#>  [51] -0.756952757  0.241751422 -1.403953767 -0.857843438  0.680878516
#>  [56]  0.386717490 -0.017351591  0.742628998  0.122158832  0.464960495
#>  [61]  0.583173121  0.908634725  1.187806790  0.281681625  0.409972221
#>  [66] -0.377847579  1.512293051  0.305171657  0.783658887 -1.002398433
#>  [71]  0.154517932 -0.357176388 -0.886066561  1.722138057  1.548826542
#>  [76] -1.997303984 -1.250570822  0.243245442 -0.062933774  0.433954306
#>  [81]  0.330368843  2.014090990  1.334869510  0.289405385 -1.911123690
#>  [86]  0.207172110 -0.932657425  1.180645865 -0.150137100  0.690575370
#>  [91]  0.828056940 -1.582978576 -0.666862907  1.435314867  0.488300066
#>  [96] -0.024997797 -0.100409758 -0.358671367 -0.840410777 -0.704907535
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 

# Define some variables in the call and take the rest from the environment
nlist(a, b, veggies = c("lettuce", "spinach"), fruits = c("banana", "papaya"))

#> $a
#>   [1]  0.632021803 -0.691559915 -0.698551147 -2.126358177 -0.189012752
#>   [6]  0.205477737 -1.338316688 -1.155834491  0.185828351 -0.953053705
#>  [11]  1.035186689 -0.380511294 -0.804927356 -0.278846749  0.283314615
#>  [16]  1.065870044  0.284305792  1.499735240 -1.193944998 -0.299988252
#>  [21]  0.423484063 -0.475867734  2.043626802  0.008833826 -0.597469553
#>  [26]  0.893984469  0.145662593  0.159455138 -2.349083183  0.864750814
#>  [31] -0.044704628  1.108497705  0.365317198  0.067778239 -2.258293294
#>  [36] -0.307612797 -1.608490159 -1.172991402  1.116213939 -0.451311914
#>  [41]  1.139423503  0.147087735  0.223140303 -1.797794624 -0.257989149
#>  [46] -1.456310534  0.553609616 -1.424251381  1.181560514  0.070606527
#>  [51] -0.756952757  0.241751422 -1.403953767 -0.857843438  0.680878516
#>  [56]  0.386717490 -0.017351591  0.742628998  0.122158832  0.464960495
#>  [61]  0.583173121  0.908634725  1.187806790  0.281681625  0.409972221
#>  [66] -0.377847579  1.512293051  0.305171657  0.783658887 -1.002398433
#>  [71]  0.154517932 -0.357176388 -0.886066561  1.722138057  1.548826542
#>  [76] -1.997303984 -1.250570822  0.243245442 -0.062933774  0.433954306
#>  [81]  0.330368843  2.014090990  1.334869510  0.289405385 -1.911123690
#>  [86]  0.207172110 -0.932657425  1.180645865 -0.150137100  0.690575370
#>  [91]  0.828056940 -1.582978576 -0.666862907  1.435314867  0.488300066
#>  [96] -0.024997797 -0.100409758 -0.358671367 -0.840410777 -0.704907535
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 
#> $veggies
#> [1] "lettuce" "spinach"
#> 
#> $fruits
#> [1] "banana" "papaya"
#>