Named lists

Create a named list using specified names or, if names are omitted, using the names of the objects in the list. The code list(a = a, b = b) becomes nlist(a,b) and list(a = a, b = 2) becomes nlist(a, b = 2), etc.

nlist(...)

Arguments

...

Objects to include in the list.

Value

A named list.

Examples

# All variables already defined
a <- rnorm(100)
b <- mat.or.vec(10, 3)
nlist(a,b)
#> $a
#>   [1] -1.28201735  0.43491262  0.62775584  0.26845933 -0.08916891  0.86577881
#>   [7]  2.43727729 -0.99105513 -0.34274484  0.15760204 -0.34590829  0.70710961
#>  [13] -0.34420784  0.35122361 -1.36091708 -0.88963630 -0.57786186  0.39287853
#>  [19] -0.71130236  0.01273766  0.12562046 -0.34968048  0.37516215 -1.73575212
#>  [25] -0.05117519  0.16517627 -0.05796315  1.32253967  0.07539235  1.57717896
#>  [31]  1.49641035 -0.51403594  0.06457554 -0.14871888  0.25444098  0.76371675
#>  [37]  0.91323798  1.14091338 -0.60698706  0.75772944 -0.67794823  1.00360049
#>  [43]  0.19255242  0.78132373 -0.62046663 -0.17532368 -0.73710668  0.65080196
#>  [49] -1.21428472 -0.14812666 -1.27851334  0.83947802 -0.05206098  0.10326899
#>  [55]  1.59880177  0.48041613  0.55295331  0.68561668 -0.91756503  0.32984826
#>  [61]  1.78893795 -1.23286732 -1.20310721  0.36671052  0.33381600  1.62112306
#>  [67]  0.03764495 -0.09797660 -0.16622660 -0.61587031  1.46530802  0.07384489
#>  [73] -0.21874320 -0.48893814 -1.41118628 -0.50967348 -0.61549183 -1.78277171
#>  [79]  1.13210577 -0.39532909 -0.53921150  2.18315778  1.11052189 -0.11300859
#>  [85] -1.45921256 -1.38248303 -0.97706036 -1.28281166  0.88960610  1.07551157
#>  [91]  0.87165853  1.74091654  0.61500387  0.33595030 -0.68544156  0.90224105
#>  [97]  1.70501137 -0.11119136  0.19051851 -0.10950033
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 

# Define some variables in the call and take the rest from the environment
nlist(a, b, veggies = c("lettuce", "spinach"), fruits = c("banana", "papaya"))
#> $a
#>   [1] -1.28201735  0.43491262  0.62775584  0.26845933 -0.08916891  0.86577881
#>   [7]  2.43727729 -0.99105513 -0.34274484  0.15760204 -0.34590829  0.70710961
#>  [13] -0.34420784  0.35122361 -1.36091708 -0.88963630 -0.57786186  0.39287853
#>  [19] -0.71130236  0.01273766  0.12562046 -0.34968048  0.37516215 -1.73575212
#>  [25] -0.05117519  0.16517627 -0.05796315  1.32253967  0.07539235  1.57717896
#>  [31]  1.49641035 -0.51403594  0.06457554 -0.14871888  0.25444098  0.76371675
#>  [37]  0.91323798  1.14091338 -0.60698706  0.75772944 -0.67794823  1.00360049
#>  [43]  0.19255242  0.78132373 -0.62046663 -0.17532368 -0.73710668  0.65080196
#>  [49] -1.21428472 -0.14812666 -1.27851334  0.83947802 -0.05206098  0.10326899
#>  [55]  1.59880177  0.48041613  0.55295331  0.68561668 -0.91756503  0.32984826
#>  [61]  1.78893795 -1.23286732 -1.20310721  0.36671052  0.33381600  1.62112306
#>  [67]  0.03764495 -0.09797660 -0.16622660 -0.61587031  1.46530802  0.07384489
#>  [73] -0.21874320 -0.48893814 -1.41118628 -0.50967348 -0.61549183 -1.78277171
#>  [79]  1.13210577 -0.39532909 -0.53921150  2.18315778  1.11052189 -0.11300859
#>  [85] -1.45921256 -1.38248303 -0.97706036 -1.28281166  0.88960610  1.07551157
#>  [91]  0.87165853  1.74091654  0.61500387  0.33595030 -0.68544156  0.90224105
#>  [97]  1.70501137 -0.11119136  0.19051851 -0.10950033
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 
#> $veggies
#> [1] "lettuce" "spinach"
#> 
#> $fruits
#> [1] "banana" "papaya"
#>