Named lists

Create a named list using specified names or, if names are omitted, using the names of the objects in the list. The code list(a = a, b = b) becomes nlist(a,b) and list(a = a, b = 2) becomes nlist(a, b = 2), etc.

nlist(...)

Arguments

...

Objects to include in the list.

Value

A named list.

Examples

# All variables already defined
a <- rnorm(100)
b <- mat.or.vec(10, 3)
nlist(a,b)
#> $a
#>   [1]  0.58303726 -1.18005670  2.02906483  0.04563914  1.18767113  0.13277372
#>   [7]  0.73074005  0.27421791 -0.98071223  0.89616912  0.09235610  1.42260024
#>  [13]  1.63958660 -0.71554286 -1.72766531 -0.26474932  0.46402132  0.18328496
#>  [19]  0.37014325 -0.46539846 -1.35699064 -0.94756588 -1.32675628  1.71944404
#>  [25] -0.87898014 -0.02435987 -0.35046913 -1.49272159 -1.34795616 -0.87239734
#>  [31]  0.36550662 -0.95172768 -0.20884901 -0.58875870 -0.71210362  2.55268532
#>  [37] -0.15015995  0.49893157  0.48499617 -0.46047212  0.47671215  1.44219503
#>  [43] -0.53252713 -1.00355596 -0.79759812 -0.35212601 -0.12575206 -1.12890042
#>  [49]  1.81283778  0.54590031  0.27614100 -1.78703622 -0.69897022  0.66958072
#>  [55]  1.16590622 -0.10045622 -0.19541553  1.01477218 -0.29486217  1.95328836
#>  [61]  0.22345390 -0.89055374 -0.21735722  0.64246743 -0.19687145 -0.92607721
#>  [67]  1.18453783  0.68133067 -0.22411617 -0.42614768  1.33986614 -0.72595917
#>  [73]  0.54091251  0.71569348  0.08556613 -0.98998589  0.27201072 -2.00633480
#>  [79] -2.45095243  0.04180641  1.27960696  0.32538273 -0.45765369  0.78471577
#>  [85] -0.03627547  1.09984716 -0.93415634  1.16878532  0.71871116  0.68790149
#>  [91]  0.28979210  0.69836408  1.10855493  0.27345095  0.98693726 -0.58554379
#>  [97] -0.01686564  0.88033592 -0.50875613  0.33882608
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 

# Define some variables in the call and take the rest from the environment
nlist(a, b, veggies = c("lettuce", "spinach"), fruits = c("banana", "papaya"))
#> $a
#>   [1]  0.58303726 -1.18005670  2.02906483  0.04563914  1.18767113  0.13277372
#>   [7]  0.73074005  0.27421791 -0.98071223  0.89616912  0.09235610  1.42260024
#>  [13]  1.63958660 -0.71554286 -1.72766531 -0.26474932  0.46402132  0.18328496
#>  [19]  0.37014325 -0.46539846 -1.35699064 -0.94756588 -1.32675628  1.71944404
#>  [25] -0.87898014 -0.02435987 -0.35046913 -1.49272159 -1.34795616 -0.87239734
#>  [31]  0.36550662 -0.95172768 -0.20884901 -0.58875870 -0.71210362  2.55268532
#>  [37] -0.15015995  0.49893157  0.48499617 -0.46047212  0.47671215  1.44219503
#>  [43] -0.53252713 -1.00355596 -0.79759812 -0.35212601 -0.12575206 -1.12890042
#>  [49]  1.81283778  0.54590031  0.27614100 -1.78703622 -0.69897022  0.66958072
#>  [55]  1.16590622 -0.10045622 -0.19541553  1.01477218 -0.29486217  1.95328836
#>  [61]  0.22345390 -0.89055374 -0.21735722  0.64246743 -0.19687145 -0.92607721
#>  [67]  1.18453783  0.68133067 -0.22411617 -0.42614768  1.33986614 -0.72595917
#>  [73]  0.54091251  0.71569348  0.08556613 -0.98998589  0.27201072 -2.00633480
#>  [79] -2.45095243  0.04180641  1.27960696  0.32538273 -0.45765369  0.78471577
#>  [85] -0.03627547  1.09984716 -0.93415634  1.16878532  0.71871116  0.68790149
#>  [91]  0.28979210  0.69836408  1.10855493  0.27345095  0.98693726 -0.58554379
#>  [97] -0.01686564  0.88033592 -0.50875613  0.33882608
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 
#> $veggies
#> [1] "lettuce" "spinach"
#> 
#> $fruits
#> [1] "banana" "papaya"
#>