Named lists

Create a named list using specified names or, if names are omitted, using the names of the objects in the list. The code list(a = a, b = b) becomes nlist(a,b) and list(a = a, b = 2) becomes nlist(a, b = 2), etc.

nlist(...)

Arguments

...
Objects to include in the list.

Value

A named list.

Examples

# All variables already defined
a <- rnorm(100)
b <- mat.or.vec(10, 3)
nlist(a,b)

#> $a
#>   [1]  1.24485982  0.45987615  0.16770583 -0.27500302  0.49232262  1.33975457
#>   [7] -0.47674422 -1.49679554 -1.20142397 -0.40746103 -0.94774810  0.85166640
#>  [13] -0.58843101 -1.36247310 -0.44963794 -0.64332503 -0.51074555  0.25720201
#>  [19] -0.41955545 -0.95222766  0.51239341 -1.74426769 -0.49751241  0.76558578
#>  [25]  0.50139307  0.33191117 -0.09107573 -0.65668203 -1.16084814  1.65593193
#>  [31] -0.74921055  0.62192056 -1.47180517  0.79819546  0.27946251  0.65787426
#>  [37]  0.59372632 -0.57792651  0.47606273 -1.18763362  0.71768805  0.13469724
#>  [43]  1.42208034  0.04073091 -0.33840975 -0.55811539  1.02781348 -0.22547048
#>  [49] -1.23163005 -0.75824474  1.08551815 -1.04909888  0.29039955 -0.44757343
#>  [55] -0.21886077 -1.60544351  0.07483969 -0.52191870 -0.06490237  1.05136997
#>  [61] -1.74215826  1.73254803  1.63352915 -0.37664970 -0.16623822 -0.03215228
#>  [67]  0.51315346  1.26822734 -1.06028265  0.61966262 -0.81962820 -0.48774264
#>  [73] -0.09108544  0.30792735 -0.37488397 -0.78785668 -1.01633952 -2.62501884
#>  [79] -0.22983211 -0.26591736  0.36381675  0.11568626  0.75730662  1.51531287
#>  [85] -2.65978385 -0.68818004 -0.42657865 -0.42620544  0.14830328  2.52780944
#>  [91] -0.14340009  1.03306764 -0.35786489  0.75772181  2.29874270 -0.07557010
#>  [97]  2.86124846 -1.40879003  0.07643776 -1.77320910
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 

# Define some variables in the call and take the rest from the environment
nlist(a, b, veggies = c("lettuce", "spinach"), fruits = c("banana", "papaya"))

#> $a
#>   [1]  1.24485982  0.45987615  0.16770583 -0.27500302  0.49232262  1.33975457
#>   [7] -0.47674422 -1.49679554 -1.20142397 -0.40746103 -0.94774810  0.85166640
#>  [13] -0.58843101 -1.36247310 -0.44963794 -0.64332503 -0.51074555  0.25720201
#>  [19] -0.41955545 -0.95222766  0.51239341 -1.74426769 -0.49751241  0.76558578
#>  [25]  0.50139307  0.33191117 -0.09107573 -0.65668203 -1.16084814  1.65593193
#>  [31] -0.74921055  0.62192056 -1.47180517  0.79819546  0.27946251  0.65787426
#>  [37]  0.59372632 -0.57792651  0.47606273 -1.18763362  0.71768805  0.13469724
#>  [43]  1.42208034  0.04073091 -0.33840975 -0.55811539  1.02781348 -0.22547048
#>  [49] -1.23163005 -0.75824474  1.08551815 -1.04909888  0.29039955 -0.44757343
#>  [55] -0.21886077 -1.60544351  0.07483969 -0.52191870 -0.06490237  1.05136997
#>  [61] -1.74215826  1.73254803  1.63352915 -0.37664970 -0.16623822 -0.03215228
#>  [67]  0.51315346  1.26822734 -1.06028265  0.61966262 -0.81962820 -0.48774264
#>  [73] -0.09108544  0.30792735 -0.37488397 -0.78785668 -1.01633952 -2.62501884
#>  [79] -0.22983211 -0.26591736  0.36381675  0.11568626  0.75730662  1.51531287
#>  [85] -2.65978385 -0.68818004 -0.42657865 -0.42620544  0.14830328  2.52780944
#>  [91] -0.14340009  1.03306764 -0.35786489  0.75772181  2.29874270 -0.07557010
#>  [97]  2.86124846 -1.40879003  0.07643776 -1.77320910
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 
#> $veggies
#> [1] "lettuce" "spinach"
#> 
#> $fruits
#> [1] "banana" "papaya"
#>