Named lists

Create a named list using specified names or, if names are omitted, using the names of the objects in the list. The code list(a = a, b = b) becomes nlist(a,b) and list(a = a, b = 2) becomes nlist(a, b = 2), etc.

nlist(...)

Arguments

...
Objects to include in the list.

Value

A named list.

See also

list

Examples

# All variables already defined
a <- rnorm(100)
b <- mat.or.vec(10, 3)
nlist(a,b)

#> $a
#>   [1]  0.3753640358  0.6671638718 -0.0085762390  0.0924404534  0.3202826420
#>   [6]  0.1650912505 -1.0532067343  0.4473946768  1.0571368249 -1.4751573301
#>  [11]  0.2007328598 -0.0704526903 -0.7503191563 -0.0940137619  0.4127494623
#>  [16] -1.2574789605  0.1970634927 -1.8923149164  0.2521761638 -0.6478509158
#>  [21] -0.2660576830 -0.2867678201 -0.5292874552 -0.5570450259 -0.5484389126
#>  [26]  0.0637433919 -1.1780599708 -1.3749018547  0.3333497355  0.5597446727
#>  [31] -1.1577567141  1.3595626446  1.3923633348 -1.6319603432  0.9944334727
#>  [36]  1.5975854685 -0.6759390641  0.5056522754  0.8613147025  0.7455505896
#>  [41]  0.9483998776 -0.2782412747  1.1369703696  0.9784894267  1.6473931433
#>  [46] -1.5064649019 -1.7938317434 -1.3929985639  0.4638976694  0.7481313731
#>  [51]  1.2235672736  0.3287331376  1.2029108012 -1.1780358319  0.9161471125
#>  [56] -0.3907389634 -0.3312925958 -0.8074691521  0.1419367987  0.3662676984
#>  [61] -2.0921697182 -0.0206317239 -0.9036675123  2.1467501026  0.5278421596
#>  [66] -1.1615360222  0.3266761594  0.4316068917  1.2290550882 -0.6725782514
#>  [71]  1.6159861082 -2.8560617851 -0.5694742942 -0.7325862311  1.6786160405
#>  [76]  0.3883841608 -0.3991853599 -1.2244360744  0.6694015824  0.9747660832
#>  [81] -0.8596496125  0.0633182687 -0.4065883075  0.2482428663 -0.2514846631
#>  [86] -0.9799973963  0.5890418888 -1.4048948877 -1.5097889155 -0.4215444125
#>  [91]  1.2801836193 -0.3640600647 -0.8525419120 -0.1058764092  1.3290976925
#>  [96] -0.2796422221  0.0507539340 -0.0793683132  0.2210828405  0.0009076064
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 

# Define some variables in the call and take the rest from the environment
nlist(a, b, veggies = c("lettuce", "spinach"), fruits = c("banana", "papaya"))

#> $a
#>   [1]  0.3753640358  0.6671638718 -0.0085762390  0.0924404534  0.3202826420
#>   [6]  0.1650912505 -1.0532067343  0.4473946768  1.0571368249 -1.4751573301
#>  [11]  0.2007328598 -0.0704526903 -0.7503191563 -0.0940137619  0.4127494623
#>  [16] -1.2574789605  0.1970634927 -1.8923149164  0.2521761638 -0.6478509158
#>  [21] -0.2660576830 -0.2867678201 -0.5292874552 -0.5570450259 -0.5484389126
#>  [26]  0.0637433919 -1.1780599708 -1.3749018547  0.3333497355  0.5597446727
#>  [31] -1.1577567141  1.3595626446  1.3923633348 -1.6319603432  0.9944334727
#>  [36]  1.5975854685 -0.6759390641  0.5056522754  0.8613147025  0.7455505896
#>  [41]  0.9483998776 -0.2782412747  1.1369703696  0.9784894267  1.6473931433
#>  [46] -1.5064649019 -1.7938317434 -1.3929985639  0.4638976694  0.7481313731
#>  [51]  1.2235672736  0.3287331376  1.2029108012 -1.1780358319  0.9161471125
#>  [56] -0.3907389634 -0.3312925958 -0.8074691521  0.1419367987  0.3662676984
#>  [61] -2.0921697182 -0.0206317239 -0.9036675123  2.1467501026  0.5278421596
#>  [66] -1.1615360222  0.3266761594  0.4316068917  1.2290550882 -0.6725782514
#>  [71]  1.6159861082 -2.8560617851 -0.5694742942 -0.7325862311  1.6786160405
#>  [76]  0.3883841608 -0.3991853599 -1.2244360744  0.6694015824  0.9747660832
#>  [81] -0.8596496125  0.0633182687 -0.4065883075  0.2482428663 -0.2514846631
#>  [86] -0.9799973963  0.5890418888 -1.4048948877 -1.5097889155 -0.4215444125
#>  [91]  1.2801836193 -0.3640600647 -0.8525419120 -0.1058764092  1.3290976925
#>  [96] -0.2796422221  0.0507539340 -0.0793683132  0.2210828405  0.0009076064
#> 
#> $b
#>       [,1] [,2] [,3]
#>  [1,]    0    0    0
#>  [2,]    0    0    0
#>  [3,]    0    0    0
#>  [4,]    0    0    0
#>  [5,]    0    0    0
#>  [6,]    0    0    0
#>  [7,]    0    0    0
#>  [8,]    0    0    0
#>  [9,]    0    0    0
#> [10,]    0    0    0
#> 
#> $veggies
#> [1] "lettuce" "spinach"
#> 
#> $fruits
#> [1] "banana" "papaya"
#>